requirements.txt
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
